+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/dispenser/ThrowableBricksBehavior.java
@@ -1,21 +1,13 @@
 package net.mehvahdjukaar.supplementaries.common.block.dispenser;
 
-import net.mehvahdjukaar.moonlight.api.util.DispenserHelper;
 import net.mehvahdjukaar.supplementaries.common.entities.ThrowableBrickEntity;
 import net.mehvahdjukaar.supplementaries.reg.ModSounds;
-import net.minecraft.core.BlockPos;
 import net.minecraft.core.BlockSource;
-import net.minecraft.core.Direction;
 import net.minecraft.core.Position;
 import net.minecraft.sounds.SoundEvent;
-import net.minecraft.sounds.SoundEvents;
-import net.minecraft.sounds.SoundSource;
-import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.DispenserBlock;
 
 class ThrowableBricksBehavior extends ProjectileBehavior {