+++ changelog.md
@@ -31,6 +31,7 @@
 - fixed quill maps not supporting vanilla decorations properly
 - new present sounds
 - changed slingshot particle
+- you can now have a 2 faced clock block
 - wind vanes compat for wilder wild
 - improved dispenser projectiles behaviors (slimeballs ender pears and bricks) to fail when a solid face block is placed against the dispenser, falling back on existing behaviors
 - changed many sounds to have their own sound event with their own subtitles